mem fixed

+++ libft/src/ft_memalloc.c
@@ -6,7 +6,7 @@
 /*   By: mzhurba <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/27 18:32:34 by mzhurba           #+#    #+#             */
-/*   Updated: 2018/11/08 16:46:36 by mzhurba          ###   ########.fr       */
+/*   Updated: 2019/09/15 18:18:29 by msaliuta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	*ft_memalloc(size_t size)
 {
-	size_t	i;
+	int		i;
 	char	*ptr;
 
 	i = -1;
 	ptr = (char*)malloc(size);
 	if (ptr)
-		while (++i < size)
+		while (++i < (int)size)
 			ptr[i] = 0;
 	else
 		return (NULL);
